Feat. 유저 . rebase

+++ src/main/java/com/hh99/hh5concertreservation/user/presentation/dto/PointBalanceResponse.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PointBalanceResponse {
-    private Integer balance;
+    private Long balance;
     
     public PointBalanceResponse(PointResult result) {
         this.balance = result.getBalance();
